refactor(tree): Separate related tree node models into smaller files
